refactor: add useUsers hook
