+++ src/interfaces/IUser.ts
@@ -1,0 +1,5 @@
+export interface IUser {
+  id: string;
+  displayName?: string;
+  score?: number;
+}
